More experimentation.
This version sets the user LED (pin 13) whenever button 0 is pressed,
and that works OK.  The TX LED flickers when things happen, so
presumably the arduino is trying to talk to the bluetooth modem; but
I'm not seeing anything on the phone's bluetooth terminal.  For the
next version, I might try outputting diagnostics to the development
host computer, via the USB connection.